backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.105.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
